move cross-val to its own module
