+++ skore/src/skore/sklearn/__init__.py
@@ -1,6 +1,6 @@
 """Enhance `sklearn` functions."""
 
-from skore.sklearn.cross_validation_reporter import CrossValidationReporter
+from skore.sklearn.cross_validation import CrossValidationReporter
 from skore.sklearn.train_test_split.train_test_split import train_test_split
 
 __all__ = [
