+++ skore/src/skore/sklearn/cross_validation/__init__.py
@@ -1,0 +1,7 @@
+"""Enhance cross-validation."""
+
+from .cross_validation_reporter import CrossValidationReporter
+
+__all__ = [
+    "CrossValidationReporter",
+]
