+++ skore/src/skore/sklearn/cross_validation/cross_validation_reporter.py
